tarea 1 y 2 completa
