+++ .gitignore
@@ -1,1 +1,3 @@
 dags/__pycache__
+
+data_set
