+++ bigdata/data_group_E/top_10_fechas_con_mas_post.py
@@ -1,12 +1,9 @@
 """
 
 Utilizar MapReduce para el grupo de datos E
+
 * Top 10 fechas con mayor cantidad de post creados
 
-* Relación entre cantidad de respuestas y sus visitas.
-
-* Del ranking de los primeros 0-100 por score, tomar el tiempo de respuesta promedio 
-e informar un único valor.
 """
 
 from functools import reduce
